modification du code au niveau des entité ajout de methode dans etudiant et aussi dans d'autre et aussi regler quelque problème (nb: ne fait pas attention sur au niveau de etudiantcontroller je fais juste des test a partir de labas ) ensuite la phase de test n'est pas terminé je dois encore vérifier que tout marche

+++ src/main/java/com/dembaandousmane/gest_priere/repository/EtudiantRepository.java
@@ -7,6 +7,6 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface EtudiantRepository extends JpaRepository<Etudiant, Integer> {
+public interface EtudiantRepository extends JpaRepository<Etudiant, Long> {
     Optional<Etudiant> findByEmail(String email);
 }
