+++ src/main/java/com/dembaandousmane/gest_priere/repository/SalleRepository.java
@@ -7,6 +7,6 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface SalleRepository extends JpaRepository<Salle, Integer> {
+public interface SalleRepository extends JpaRepository<Salle, Long> {
     List<Salle> findByStatut(String statut);
 }
